requirements.txt
psycopg2-binary==2.8.5
pyodbc==4.0.30
celery==4.4.7
sqlalchemy==1.3.18
redis==3.5.3
flask==1.1.2
gunicorn==20.0.4
slackclient==2.7.3
loguru==0.5.1